[add & update] DetailWindowへ課題追加機能の追加(未完成)
[update] DetailWindow
DetailWindowへInputAssignmentを用いた課題追加機能を追加
(ただし、現在未完成)

[add & update] InputAssignment
イベント設定機能追加
一部メンバ関数の仕様を変更
メンバ関数を追加
また、定義をinput_assignment.pyに移行

[add & update] Assignment
__init__関数の仕様を変更(既存の利用に影響なし)
__str__関数を追加

[update] Subject
add_asg関数の引数を変更
